Fix build error in cluster code.
git-svn-id: https://svn.eu.apache.org/repos/asf/incubator/qpid/trunk@691456 13f79535-47bb-0310-9956-ffa450edef68

+++ qpid/cpp/src/qpid/cluster/Connection.cpp
@@ -86,7 +86,7 @@ void Connection::deliverClose () {
 // Delivery of doOutput allows us to run the real connection doOutput()
 // which stocks up the write buffers with data.
 // 
-void Connection::deliverDoOutput(size_t requested) {
+void Connection::deliverDoOutput(uint32_t requested) {
     output.deliverDoOutput(requested);
 }
